Add support for setting ssh port
Some systems run multiple ssh servers, for instance if they host SFTP servers. 
In that case its nice to be able to configure a different ssh port for rundeck to use on some EC2 instances.

Rundeck would like the ssh port to be part of the hostname like this <hostname>:<port>
Setting a sshport.selector will append that port to the hostname if its different from 22.